# server/openapi.yaml
openapi: 3.0.0
info:
  title: Product Microservice API
  version: 1.0.0
  description: API para gestionar equipos y partidos de la MLB.
servers:
  - url: http://servicio-datos:6000
    description: Servidor local de desarrollo
paths:
...
  /api/equipos/buscar:
    get:
      summary: Buscar equipos por nombre
      description: Obtener todos los equipos que coincidan con el termino buscado
      parameters:
        - name: nombre
          in: query
          required: true
          description: El nombre o término de búsqueda del equipo.
          schema:
            type: string
      responses:
        '200':
          description: Equipos encontrados que coinciden con el texto buscado.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#components/schemas/Equipo' 
        '400':
          description: Parámetro de búsqueda no proporcionado.
        '500':
          description: Error en la búsqueda de equipos.
...
  /api/partidos:
    get:
      summary: Obtener todos los partidos por la fecha
      responses:
        '200':
          description: Todos los partidos obtenidos correctamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#components/schemas/Partido' 
        '400':
          description: Fecha no proporcionada.
        '500':
          description: Error obteniendo los partidos.
...
  /api/clasificacion:
    get:
      summary: Obtener clasificacion por año.
      description: Calcular la clasificacion del año proporcionado segun los partidos jugados por los equipo.
      parameters:
        - name: temporada
          in: query
          required: true
          description: El año de la temporada.
          schema:
            type: string
      responses:
        '200':
          description: Clasificacion obtenida correctamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#components/schemas/Clasificacion' 
        '500':
          description: Error al calcular la clasificación.
...